refactor: update all pages to typescript
