+++ pages/participant/index.tsx
@@ -1,22 +1,23 @@
 import Container from "@mui/material/Container";
 import Grid from "@mui/material/Grid";
+import { GetServerSideProps } from "next";
 import { getSession } from "next-auth/client";
 import Head from "next/head";
 import { useRouter } from "next/router";
 import { Fragment, useEffect, useState } from "react";
+import Credits from "../../components/Credits";
 import GridCard from "../../components/GridCard";
 import NavBar from "../../components/NavBar";
-import { getTeams } from "../../fetchData/getTeams";
-import Credits from "../../components/Credits";
+import { getParticipants } from "../../fetchData/getParticipants";
 
 /**
  *
- * @param {Object} participantList list of teams from the database. All teams if no query is present, else subset
+ * @param {Object} participantList list of members from the database. All members if no query is present, else subset
  * @param {string[]} roleList list of all roles, used for role query search bar auto-fill
  * @param {Object} session contains whether the user is signed in or not
  */
-export default function TeamListPage({ participantList, roleList, session }) {
-  // contains subset of team objects based on name that is typed in the search bar
+export default function MembershipPage({ participantList, roleList, session }) {
+  // contains subset of member objects based on name that is typed in the search bar
   const [filteredArray, setFilteredArray] = useState(participantList);
 
   // contains the list of all names only that is used to populate the search bar auto-fill
@@ -64,7 +65,7 @@ export default function TeamListPage({ participantList, roleList, session }) {
 
   const Grids = filteredArray.map(({ id, name }, index) => {
     return (
-      <GridCard id={id} name={name} path="team" key={index}/>
+      <GridCard id={id} name={name} path="participant" key={index}/>
     );
   });
 
@@ -72,12 +73,12 @@ export default function TeamListPage({ participantList, roleList, session }) {
     <Fragment>
       <Head>
         <title>
-          Programs{fetchQuery() === "" ? "" : `: ${fetchQuery()}`} | ACM
+          Membership{fetchQuery() === "" ? "" : `: ${fetchQuery()}`} | ACM
           Leadership
         </title>
         <meta
           property="og:title"
-          content="Programs | ACM Leadership"
+          content="Membership | ACM Leadership"
           key="title"
         />
       </Head>
@@ -105,16 +106,16 @@ export default function TeamListPage({ participantList, roleList, session }) {
   );
 }
 
-export async function getServerSideProps(context) {
-  const { teams, programs } = await getTeams(context.query.q);
+export const getServerSideProps: GetServerSideProps = async (context) => {
+  const { participants, programs } = await getParticipants(context.query.q);
   const session = await getSession(context);
 
-  if (!teams || !programs) {
+  if (!participants || !programs) {
     return {
       notFound: true,
     };
   }
   return {
-    props: { participantList: teams, roleList: programs, session }, // will be passed to the page component as props
+    props: { participantList: participants, roleList: programs, session }, // will be passed to the page component as props
   };
 }
