+++ pages/profile/[username].tsx
@@ -7,6 +7,7 @@ import Container from "@mui/material/Container";
 import TextField from "@mui/material/TextField";
 import Typography from "@mui/material/Typography";
 import axios from "axios";
+import { GetServerSideProps } from 'next';
 import { getSession } from "next-auth/client";
 import dynamic from "next/dynamic";
 import Head from "next/head";
@@ -188,7 +189,7 @@ export default function ProfilePage({ data, session }) {
   );
 }
 
-export const getServerSideProps = async (context) => {
+export const getServerSideProps: GetServerSideProps = async (context) => {
   const { username } = context.params;
   const session = await getSession(context);
